backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.19
python-pptx==1.0.2
httpx==0.27.2
pydantic==2.10.3
pydantic-ai-slim[openai]==1.73.0
aiosqlite==0.20.0
pypdf>=5,<7
